+++ test/mysql2_test.rb
@@ -130,7 +130,6 @@ describe "mysql2 connection" do
 
   it "correctly handles ActiveRecord's local timezone setting" do
     ActiveRecord::Base.default_timezone = :local
-    @db.timezone = :local
 
     time = Time.new(2020, 4, 26, 0, 0, 0)